+++ src/AllTables/rsvp/reservation.controller.ts
@@ -13,7 +13,7 @@ export const createReservationController = async (req: Request, res: Response) =
         if (!createReservation) {
             return res.json({message: "New RSVP not created"})
         } 
-        return res.status(201).json({message: "New RSVP Created!!", newreservation})            
+        return res.status(201).json({message: "New RSVP Created!!", RSVP: createReservation})            
     } catch (error: any) {
         return res.status(500).json({error: error.message})
     }
@@ -28,7 +28,7 @@ export const getAllReservationsController = async (req: Request, res: Response) 
          if (!getAllReservations || getAllReservations.length === 0) {
             return res.status(404).json({message: "No RSVP found"});
         }
-        return res.status(200).json({data: getAllReservations});
+        return res.status(200).json({reservations: getAllReservations});
     
     } catch (error: any) {
         return res.status(500).json({error: error.message})        
@@ -45,9 +45,9 @@ export const getReservationByIdController = async (req: Request, res: Response) 
         }
         const getReservationByID = await getReservationByRSVPIDService(id);
         if (!getReservationByID) {
-            return res.status(404).json({message: "RSVP not found"});
+            return res.status(404).json({message: "Reservation not found"});
         }
-        return res.status(200).json({data: getReservationByID});
+        return res.status(200).json({reservation: getReservationByID});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
@@ -63,9 +63,9 @@ export const getReservationByEventIDController = async (req: Request, res: Respo
         }
         const getReservationByID = await getReservationByEventIDService(id);
         if (!getReservationByID) {
-            return res.status(404).json({message: "RSVP not found"});
+            return res.status(404).json({message: "Reservation not found"});
         }
-        return res.status(200).json({data: getReservationByID});
+        return res.status(200).json({reservation: getReservationByID});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
@@ -80,9 +80,9 @@ export const getReservationByUserIDController = async (req: Request, res: Respon
         }
         const getReservationByID = await getReservationByUserIDService(id);
         if (!getReservationByID) {
-            return res.status(404).json({message: "RSVP not found"});
+            return res.status(404).json({message: "Reservation not found"});
         }
-        return res.status(200).json({data: getReservationByID});
+        return res.status(200).json({reservation: getReservationByID});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
@@ -99,18 +99,18 @@ export const updateReservationController = async (req: Request, res: Response) =
         
         const reservationUpdates = req.body;
         
-        // Convert dueDate to Date object if provided
-        // if (carUpdates.dueDate) {
-        //     carUpdates.dueDate = new Date(carUpdates.dueDate);
-        // }
+        const rsvpExisted = await getReservationByRSVPIDService(id)
+        if(!rsvpExisted){
+            return res.status(404).json({message: "Reservation not found!!"})
+        }
 
-        const updatedMessage = await updateReservationService(id, reservationUpdates);
-        if (!updatedMessage) {
-            return res.status(404).json({message: "Reservation not found!!"});
+        const updated = await updateReservationService(id, reservationUpdates);
+        if (!updated) {
+            return res.status(404).json({message: "Reservation not updated"});
         }        
-        return res.status(200).json({message: updatedMessage});
+        return res.status(200).json({message: "Reservation updated successfully ✅", updated});
     } catch (error: any) {
-        return res.status(500).json({error: error.message});
+        return res.status(500).json({message: "Reservation not updated"});
     }
 }
 
@@ -121,14 +121,18 @@ export const deleteReservationController = async (req: Request, res: Response) =
         if (isNaN(id)) {
             return res.status(400).json({message: "Invalid ID format"});
         }
-        
-        const deletedMessage = await deleteReservationService(id);
-        if (!deletedMessage) {
-            return res.status(404).json({message: "Reservation not found !!!"});
+
+        const deletedRSVP = await getReservationByRSVPIDService(id)
+        if(!deletedRSVP){
+            return res.status(404).json({message: "Reservation not found !!!  Failed to delete"})
         }
-        return res.status(200).json({message: deletedMessage});
+        
+        const deleted = await deleteReservationService(id);
+        if (deleted.length > 0) {
+        return res.status(200).json({message: "Reservation deleted Successfully!!", deletedRSVP});
+        }
     } catch (error: any) {
-        return res.status(500).json({error: error.message});
+        return res.status(500).json({message: "Reservation not deleted!!"});
     }
 }
