+++ src/AllTables/tickets/ticket.controller.ts
@@ -1,7 +1,7 @@
 // API
 
 import { Request, Response } from "express";
-import { createTicketService, deleteTicketService, getAllTicketsService, getTicketByIDService, getTicketByUserIDService, updateTicketService } from "../venues/venue.service";
+import { createTicketService, deleteTicketService, getAllTicketsService, getTicketByIDService, getTicketByUserIDService, updateTicketService } from "./ticket.service";
 
 
 //Create a new Ticket
@@ -9,11 +9,11 @@ export const createTicketController = async (req: Request, res: Response) =>{
     try {
         const newTicket = req.body;
 
-        const createCar = await createTicketService(newTicket)
-        if (!createCar) {
+        const createTicket = await createTicketService(newTicket)
+        if (!createTicket) {
             return res.json({message: "New Ticket not created"})
         } 
-        return res.status(201).json({message: "New Ticket Created!!", newTicket})            
+        return res.status(201).json({message: "Customer support ticket created successfully", newTicket: createTicket})            
     } catch (error: any) {
         return res.status(500).json({error: error.message})
     }
@@ -25,10 +25,10 @@ export const getAllTicketController = async (req: Request, res: Response) =>{
         const allTickets =req.body;
 
         const getAllTickets = await getAllTicketsService();
-         if (!getAllTickets || getAllTickets.length === 0) {
+         if (!getAllTickets) {
             return res.status(404).json({message: "No Tickets found"});
         }
-        return res.status(200).json({Tickets: getAllTickets});
+        return res.status(200).json({message: "Customer support tickets retrieved successfully", Tickets: getAllTickets});
     
     } catch (error: any) {
         return res.status(500).json({error: error.message})        
@@ -47,13 +47,13 @@ export const getTicketByIdController = async (req: Request, res: Response) => {
         if (!getTicketByID) {
             return res.status(404).json({message: "Ticket not found"});
         }
-        return res.status(200).json({data: getTicketByID});
+        return res.status(200).json({message: "Customer support ticket retrieved successfully", Ticket: getTicketByID});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
 }
 
-// get Ticket by id controller
+// get Ticket by userid controller
 export const getTicketByUserIdController = async (req: Request, res: Response) => {
     try {
         const id  = parseInt (req.params.id);
@@ -64,7 +64,7 @@ export const getTicketByUserIdController = async (req: Request, res: Response) =
         if (!getTicketByID) {
             return res.status(404).json({message: "Ticket not found"});
         }
-        return res.status(200).json({data: getTicketByID});
+        return res.status(200).json({message: "Customer support ticket retrieved successfully", Ticket: getTicketByID});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
@@ -80,16 +80,20 @@ export const updateTicketController = async (req: Request, res: Response) => {
         
         const TicketUpdates = req.body;
         
-        // Convert dueDate to Date object if provided
-        if (TicketUpdates.dueDate) {
-            TicketUpdates.dueDate = new Date(TicketUpdates.dueDate);
-        }
+       const getEventByID = await getTicketByIDService(id);
+               if (!getEventByID) {
+                   return res.status(404).json({message: "Ticket not found"});
+               }  
 
-        const updatedMessage = await updateTicketService(id, TicketUpdates);
-        if (!updatedMessage) {
-            return res.status(404).json({message: "Ticket not found !!"});
-        }        
-        return res.status(200).json({message: updatedMessage});
+        const updatedTicket = await updateTicketService(id, TicketUpdates);
+        if (!updatedTicket) {
+            return res.status(404).json({message: "Ticket not Updated !!"});
+        }     
+           
+        return res.status(200).json({
+            message: "Customer support ticket updated successfully", 
+            updatedTicket
+        });
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
@@ -103,11 +107,17 @@ export const deleteTicketController = async (req: Request, res: Response) => {
             return res.status(400).json({message: "Invalid ID format"});
         }
         
-        const deletedMessage = await deleteTicketService(id);
-        if (!deletedMessage) {
-            return res.status(404).json({message: "Ticket not found !!!"});
-        }
-        return res.status(200).json({message: deletedMessage});
+        const deletedTicket = await getTicketByIDService(id);
+            if (!deletedTicket ){
+                return res.status(404).json({message: "Ticket not found !!!  Failed to delete"});
+            }
+            
+            const delEvent =await deleteTicketService(id);        
+    
+            if(delEvent.length >0){
+                return res.status(200).json({message: "Customer support ticket deleted successfully", deletedTicket});
+            }
+            return res.status(404).json({message: "Failed to delete"});
     } catch (error: any) {
         return res.status(500).json({error: error.message});
     }
